convert image edit things (all claude code)
