+++ app/forms/image-edit.tsx
@@ -6,64 +6,21 @@
  * Copyright Oxide Computer Company
  */
 import { useForm } from 'react-hook-form'
-import { useNavigate, type LoaderFunctionArgs } from 'react-router'
+import { useNavigate } from 'react-router'
 
-import { apiQueryClient, usePrefetchedApiQuery, type Image } from '@oxide/api'
+import { type Image } from '@oxide/api'
 import { Images16Icon } from '@oxide/design-system/icons/react'
 
 import { DescriptionField } from '~/components/form/fields/DescriptionField'
 import { NameField } from '~/components/form/fields/NameField'
 import { TextField } from '~/components/form/fields/TextField'
 import { SideModalForm } from '~/components/form/SideModalForm'
-import {
-  getProjectImageSelector,
-  getSiloImageSelector,
-  useProjectImageSelector,
-  useSiloImageSelector,
-} from '~/hooks/use-params'
 import { PropertiesTable } from '~/ui/lib/PropertiesTable'
 import { ResourceLabel } from '~/ui/lib/SideModal'
-import { pb } from '~/util/path-builder'
 import { capitalize } from '~/util/str'
 import { bytesToGiB } from '~/util/units'
 
-export const ProjectImageEdit = {
-  loader: async ({ params }: LoaderFunctionArgs) => {
-    const { project, image } = getProjectImageSelector(params)
-    await apiQueryClient.prefetchQuery('imageView', { path: { image }, query: { project } })
-    return null
-  },
-  Component: EditProjectImageSideModalForm,
-}
-
-export const SiloImageEdit = {
-  loader: async ({ params }: LoaderFunctionArgs) => {
-    const { image } = getSiloImageSelector(params)
-    await apiQueryClient.prefetchQuery('imageView', { path: { image } })
-    return null
-  },
-  Component: EditSiloImageSideModalForm,
-}
-
-function EditProjectImageSideModalForm() {
-  const { project, image } = useProjectImageSelector()
-  const { data } = usePrefetchedApiQuery('imageView', {
-    path: { image },
-    query: { project },
-  })
-
-  const dismissLink = pb.projectImages({ project })
-  return <EditImageSideModalForm image={data} dismissLink={dismissLink} type="Project" />
-}
-
-function EditSiloImageSideModalForm() {
-  const { image } = useSiloImageSelector()
-  const { data } = usePrefetchedApiQuery('imageView', { path: { image } })
-
-  return <EditImageSideModalForm image={data} dismissLink={pb.siloImages()} type="Silo" />
-}
-
-function EditImageSideModalForm({
+export function EditImageSideModalForm({
   image,
   dismissLink,
   type,
